Split clockSelector into separate UIViews
